Tutorial Part 3 & 4: Scaffolding and Seed Data

+++ RazorPagesBooks/Models/Book.cs
@@ -11,5 +11,5 @@ public class Book
 
     [DataType(DataType.Date)]
     public DateTime FirstPublished { get; set; }
-    public int IBSN { get; set; }
+    public long IBSN { get; set; }
 }
